Attendance 변경 메서드 리팩토링(상태변화 정책 application 쪽으로 이동) 및 테스트 완료

+++ back/src/test/java/com/haruon/groupware/domain/empInfo/emp/EmpTest.java
@@ -138,7 +138,7 @@ class EmpTest {
                 "image/png",
                 "원본",
                 "png",
-                ((long) 1024 * 1024)
+                (1024L * 1024)
         );
 
         assertThat(emp.getEmpFiles()).hasSize(1);
